% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/boottol.R
\name{boottol}
\alias{boottol}
\title{Creates 99 percent bootstrap percentile confidence intervals around credit scoring validation statistics}
\usage{
boottol(score, target, bootsamp, lth, grp)
}
\arguments{
\item{score}{- The score groupings.}

\item{target}{- The binary target variable.}

\item{bootsamp}{- How many bootstrap samples to be computed. When bootsamp is too low, a warning will be produced.}

\item{lth}{- If the score is rank ordered least likely to most likely then set equal to True, else False.}

\item{grp}{- An integer value of how the scores are grouped.}
}
\description{
Creates 99 percent bootstrap percentile confidence intervals around credit scoring validation statistics
}
\examples{
data("data")
\dontrun{tol.level=boottol(score=data$Score,target=data$Target,bootsamp=2000,lth="True",grp=10)}
}

