\name{data}
\alias{data}
\docType{data}
\title{
Bootstrap Tolerance Levels for Credit Scoring Validation Statistics: Rank Order Statistics
}
\description{
This data was randomly generated. The Target varible is binary. The Score variable ranges from 0-90 in intervals of 10.
}
\usage{data(data)}
\format{
  A data frame with 5000 observations on the following 3 variables.
  \describe{
    \item{\code{Id}}{a numeric vector}
    \item{\code{Target}}{a numeric vector}
    \item{\code{Score}}{a numeric vector}
  }
}
\examples{
data("data")
}
\keyword{datasets}
