\name{vasdata}
\alias{vasdata}
\docType{data}
\title{
Bootstrap Tolerance Levels for Credit Scoring Validation Statistics: Vasicek Test
}
\description{
This data was randomly generated. The Response variable is a response target and is binary. The Segment variable ranges from 0-60 in intervals of 10.
}
\usage{data(vasdata)}
\format{
  A data frame with 7000 observations on the following 2 variables.
  \describe{
    \item{\code{segment}}{a numeric vector}
    \item{\code{response}}{a numeric vector}
  }
}
\examples{
data("vasdata")
}
\keyword{datasets}
